execution-time-estimator: move aspect from skiller-simulator
Move the aspect into the separate execution-time-estimator plugin and
adapt accordingly.
