+++ src/plugins/Makefile
@@ -31,7 +31,7 @@ SUBDIRS	= bbsync bblogger webview ttmainloop rrd \
 	  openrave-robot-memory openni refboxcomm ros player xmlrpc gossip \
 	  robot_state_publisher gazebo dynamixel navgraph-interactive \
 	  pddl-planner stn-generator clips-executive \
-	  asp plexil cedar gologpp hardware-models
+	  asp plexil cedar gologpp hardware-models execution-time-estimator
 
 include $(BUILDSYSDIR)/rules.mk
 
@@ -51,7 +51,7 @@ openrave-robot-memory: robot-memory openrave
 pddl-planner: robot-memory
 gazebo: robotino
 skiller: navgraph
-skiller-simulator: skiller
+skiller-simulator: skiller execution-time-estimator
 skiller-simulator-navgraph: skiller-simulator
 perception: mongodb
 navgraph-generator: navgraph amcl
