+++ src/plugins/execution-time-estimator/execution_time_estimator_aspect/execution_time_estimator_aspect_inifin.cpp
@@ -21,7 +21,6 @@
 #include "execution_time_estimator_aspect_inifin.h"
 
 namespace fawkes {
-namespace skiller_simulator {
 
 /** @class ExecutionTimeEstimatorsAspectIniFin
  * The Aspect IniFin for the ExecutionTimeEstimatorsAspect.
@@ -72,5 +71,4 @@ ExecutionTimeEstimatorsAspectIniFin::finalize(Thread *thread)
 	get_aspect(thread)->finalize_ExecutionTimeEstimatorsAspect();
 }
 
-} // namespace skiller_simulator
 } // namespace fawkes
