+++ src/plugins/skiller-simulator/exec_thread.h
@@ -21,10 +21,6 @@
 #ifndef _PLUGINS_SKILLER_SIMULATOR_EXEC_THREAD_H_
 #define _PLUGINS_SKILLER_SIMULATOR_EXEC_THREAD_H_
 
-#include "execution_time_estimator_aspect/execution_time_estimator_aspect.h"
-#include "execution_time_estimator_aspect/execution_time_estimator_aspect_inifin.h"
-
-#include <aspect/aspect_provider.h>
 #include <aspect/blackboard.h>
 #include <aspect/blocked_timing.h>
 #include <aspect/clock.h>
@@ -32,6 +28,7 @@
 #include <aspect/logging.h>
 #include <core/threading/thread.h>
 #include <interfaces/SkillerInterface.h>
+#include <plugins/execution-time-estimator/execution_time_estimator_aspect/execution_time_estimator_aspect.h>
 #include <plugins/skiller/skiller_feature.h>
 
 class SkillerSimulatorExecutionThread : public fawkes::Thread,
@@ -39,7 +36,7 @@ class SkillerSimulatorExecutionThread : public fawkes::Thread,
                                         public fawkes::LoggingAspect,
                                         public fawkes::BlackBoardAspect,
                                         public fawkes::ConfigurableAspect,
-                                        public fawkes::AspectProviderAspect
+                                        public fawkes::ExecutionTimeEstimatorsAspect
 {
 public:
 	SkillerSimulatorExecutionThread();
@@ -61,8 +58,6 @@ private:
 	fawkes::SkillerInterface *skiller_if_;
 	float                     default_skill_runtime_;
 	fawkes::Time              skill_starttime_;
-	fawkes::skiller_simulator::ExecutionTimeEstimatorManager       execution_time_estimator_manager_;
-	fawkes::skiller_simulator::ExecutionTimeEstimatorsAspectIniFin provider_inifin_;
 };
 
 #endif /* !_PLUGINS_SKILLER_SIMULATOR_EXEC_THREAD_H_ */
